core/Makefile: avoid exceeding MAX_ARG_STRLEN
The current method of generating the apkcerts file can fail with
a "make: execvp: /bin/bash: Argument list too long" when the path
to the toplevel directory is long enough.

In my case, the MAX_ARG_STRLEN of 131072 was exceeded with a toplevel
directory length of 52 (with 427 apkcerts).

This reduces the length of the command-line string generated by not
repeating the output path for each entry (it cd's there first).

Change-Id: I9739f9aefdf5e0febed66517c7b418a4041f49f6